[GTK] LDFLAGS is ignored when running g-ir-scanner https://bugs.webkit.org/show_bug.cgi?id=138832
Patch by Ting-Wei Lan <lantw44@gmail.com> on 2014-12-09
Reviewed by Carlos Garcia Campos.

GTK+ requires libintl on FreeBSD because there is no gettext
implementation in FreeBSD libc. In order to link to libintl, which is
not located in the default search path of the linker, we need to add
the required path via the linker command line. However, LDFLAGS is
ignored when running g-ir-scanner starting from r167873, so we have
to manually pick required flags from CMAKE_SHARED_LINKER_FLAGS and
put them in correct order to prevent g-ir-scanner from linking to
the installed version of libraries.

* PlatformGTK.cmake:

Identifier: 157293@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177016 268f45cc-cd09-0410-ab3c-d52691b4dbfc