http://omfgdogs.info/ only animates when you resize the window https://bugs.webkit.org/show_bug.cgi?id=139435 <rdar://problem/19190493>
Reviewed by Simon Fraser.

Source/WebCore:

After r163928, we would fail to animate a gif if:
- it is used as a background image of a 0-height html element
- it is used as a background image of a 0-height body element whose
  background is delegated to the root (because the root has no
  background).

This is because in such cases, shouldRepaintForImageAnimation()
should use the background rect instead of the renderer's overflow
rect to determine if the image is inside the viewport. Both cases
are addressed in this patch.

Tests: fast/images/animated-gif-body-delegated-background-image.html
       fast/images/animated-gif-body-outside-viewport.html
       fast/images/animated-gif-html-background-image.html

* rendering/RenderElement.cpp:
(WebCore::shouldRepaintForImageAnimation):
* testing/Internals.cpp:
(WebCore::Internals::hasPausedImageAnimations):
* testing/Internals.h:
* testing/Internals.idl:

LayoutTests:

Add layout tests to make sure a gif image is still animated if:
- It is used as a background image of a 0-height html element
- It is used as a background image of a 0-height body element that
  is delegated to the root (because the root has no background)

Also add a layout test to make sure we still stop the gif animation
if it is used as background image of a body that is outside the
viewport.

* fast/images/animated-gif-body-delegated-background-image-expected.txt: Added.
* fast/images/animated-gif-body-delegated-background-image.html: Added.
* fast/images/animated-gif-body-outside-viewport-expected.txt: Added.
* fast/images/animated-gif-body-outside-viewport.html: Added.
* fast/images/animated-gif-html-background-image-expected.txt: Added.
* fast/images/animated-gif-html-background-image.html: Added.

Identifier: 157388@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177135 268f45cc-cd09-0410-ab3c-d52691b4dbfc