[GStreamer] Disable gst-plugin-scanner if seccomp filters are enabled https://bugs.webkit.org/show_bug.cgi?id=140069
Patch by Michael Catanzaro <mcatanzaro@igalia.com> on 2015-01-05
Reviewed by Philippe Normand.

If seccomp filters are enabled, any child processes we create are doomed
if they try to use a trapped syscall: the child will receive SIGSYS and
dump core. There appears to be no way around this.

Fortunately, the gstreamer developers had mad prescience and provided
us with a way to move plugin scanning in-process, so we don't need the
helper binary at all.

No new tests because this is covered by existing tests when seccomp
filters are enabled.

* platform/graphics/gstreamer/GStreamerUtilities.cpp:
(WebCore::initializeGStreamer):

Identifier: 158043@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177896 268f45cc-cd09-0410-ab3c-d52691b4dbfc