Remove an old workaround for an old unsupported plug-in https://bugs.webkit.org/show_bug.cgi?id=140141
Reviewed by Brent Fulgham.

The old, pre-Vista(!) plug-in is no longer available so we don't need to have this workaround in place.
(If for some bizarre reason we'd ever want to bring it back it should be a flag that the plug-in code sets on the timer,
the shared timer code shouldn't know about plug-ins at all).

* platform/win/SharedTimerWin.cpp:
(WebCore::TimerWindowWndProc):

Identifier: 158110@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177991 268f45cc-cd09-0410-ab3c-d52691b4dbfc