Out of bounds read in IdentifierArena::makeIdentifier https://bugs.webkit.org/show_bug.cgi?id=140376
Patch by Alexey Proskuryakov.

Reviewed and ChangeLogged by Geoffrey Garen.

No test, since this is a small past-the-end read, which is very
difficult to turn into a reproducible failing test -- and existing tests
crash reliably using ASan.

* parser/ParserArena.h:
(JSC::IdentifierArena::makeIdentifier):
(JSC::IdentifierArena::makeIdentifierLCharFromUChar): Check for a
zero-length string input, like we do in the literal parser, since it is
not valid to dereference characters in a zero-length string.

A zero-length string is allowed in JavaScript -- for example, "".


Identifier: 158362@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178311 268f45cc-cd09-0410-ab3c-d52691b4dbfc