[GTK] Do not schedule a redraw for the redirected XComposite window when not in accelerated compositing mode https://bugs.webkit.org/show_bug.cgi?id=140354
Reviewed by Martin Robinson.

The RedirectedXCompositeWindow uses XDamage to be notified when
the XWindow has been modified. Since the redirected window is
created unconditionally we always receive a XDamage notify when
the window is mapped, and we schedule a redraw in the web view
even when the web view is not in accelerated compositing mode.

* UIProcess/API/gtk/WebKitWebViewBase.cpp:
(webkitWebViewBaseRealize): Only call gtk_widget_queue_draw() when
in accelerated compositing mode.

Identifier: 158420@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178415 268f45cc-cd09-0410-ab3c-d52691b4dbfc